integration tests: move regression tests to a separate file
